@@ -17,13 +17,14 @@
         }
     ],
     "require": {
-        "php": ">=7.0",
+        "php": ">=7.4",
         "overtrue/http": "1.1.3",
         "pimple/pimple": "^3.0",
         "psr/simple-cache": "^1.0",
         "monolog/monolog": "^1.23 || ^2.0",
         "symfony/cache": "^3.3 || ^4.0 || ^5.0",
-        "symfony/http-foundation": "^3.2 || ^4.0 || ^5.0"
+        "symfony/http-foundation": "^3.2 || ^4.0 || ^5.0",
+        "ext-json": "*"
     },
     "require-dev": {
         "mockery/mockery": "^1.0",
